vacuum_forge/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "vacuumforge"
version = "0.1.0"
description = "Symbolic research workbench for vacuum-based gravity theory"
requires-python = ">=3.11"
dependencies = [
    "sympy>=1.12",
    "networkx>=3.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "jinja2>=3.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]

[project.scripts]
vacuumforge = "vacuumforge.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
